Change design to cart, implement confirm to empty cart

+++ resources/js/app.js
@@ -203,13 +203,15 @@ const app = new Vue({
             localStorage.setItem('total_price', JSON.stringify(this.total_price));
         },
         emptyCart() {
-            this.cart.forEach((dish, ind) => {
-                this.cart[ind].quantity = 1; //dish
-            })
-            this.cart = [];
-            this.total_price = 0;
-            localStorage.setItem('cart', JSON.stringify(this.cart));
-            localStorage.setItem('total_price', JSON.stringify(this.total_price));
+            if (confirm('Confermi di voler svuotare il carrello?')) {
+                this.cart.forEach((dish, ind) => {
+                    this.cart[ind].quantity = 1; //dish
+                })
+                this.cart = [];
+                this.total_price = 0;
+                localStorage.setItem('cart', JSON.stringify(this.cart));
+                localStorage.setItem('total_price', JSON.stringify(this.total_price));
+            }
         },
         resetCategories() {
             this.selectedInApi = '';
